Move types to client crate.
