+++ integration_tests/tests/integration/protocols/frost_ed25519/sign.rs
@@ -1,14 +1,15 @@
 use super::dkg::run_keygen;
 use super::make_signing_message;
 use anyhow::Result;
-use mpc_client::frost::ed25519::sign;
+use mpc_client::{
+    frost::ed25519::sign, ServerOptions, SessionOptions,
+};
 use mpc_driver::{
     frost::ed25519::{
         ed25519_dalek::{SigningKey, VerifyingKey},
         KeyShare, Participant, PartyOptions,
     },
     frost_ed25519::{keys, Identifier},
-    ServerOptions, SessionOptions,
 };
 use mpc_protocol::{generate_keypair, Keypair, Parameters};
 use std::collections::BTreeMap;
